+++ src/modules/GoPro/GoProStream.py
@@ -44,20 +44,24 @@ SAVE_FILENAME="goprofeed3"
 SAVE_FORMAT="ts"
 SAVE_LOCATION="/tmp/"
 # destination rtmp application to determine text sent to Tweet
-LOGLEVEL = "debug"
-DESTINATION = "shower"
+LOGLEVEL="debug"
+DESTINATION="shower"
+MODE="remote"
 
 # arguments
 i = 0
 while i < len(sys.argv):
-    if '-d' in str(sys.argv[i]):
+    if '-destination' in str(sys.argv[i]):
         DESTINATION = str(sys.argv[i+1])
     if '-loglevel' in str(sys.argv[i]):
         LOGLEVEL = str(sys.argv[i+1])
+    if '-mode' in str(sys.argv[i]):
+        MODE = str(sys.argv[i+1])
     i += 1
 
 print("LOGLEVEL: "+LOGLEVEL)
 print("DESTINATION: "+DESTINATION)
+print("MODE: "+MODE)
 
 print("Connecting to GoPro Media...")
 sys.stdout.flush()
@@ -124,10 +128,10 @@ def gopro_live():
 			print("Recording locally: " + str(SAVE))
 			print("Recording stored in: " + SAVELOCATION)
 			print("Note: Preview is not available when saving the stream.")
-			#subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -probesize 100M -analyzeduration 100M -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://76.91.217.22:1935/"+DESTINATION, shell=True)
-			#subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -movflags faststart -probesize 100M -analyzeduration 100M -preset slow -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://alexdeeznuts.com:1935/"+DESTINATION, shell=True)
-			# subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -movflags faststart -probesize 100M -analyzeduration 100M -preset slow -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://127.0.0.1:8935/live/stream?sign=1558308292123-82cade5510d1a06dac14c93c03f8ddce", shell=True)
-			subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -movflags faststart -probesize 100M -analyzeduration 100M -preset slow -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://127.0.0.1:1935/shower", shell=True)
+			if str(MODE)=="remote"
+				subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -movflags faststart -probesize 100M -analyzeduration 100M -preset slow -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://alexdeeznuts.com:1935/"+DESTINATION, shell=True)
+			else:
+				subprocess.Popen("ffmpeg -re -i 'udp://10.5.5.100:8554' -loglevel "+LOGLEVEL+" -movflags faststart -probesize 100M -analyzeduration 100M -preset slow -f:v mpegts -acodec copy -vcodec copy -f flv rtmp://127.0.0.1:1935/"+DESTINATION, shell=True)
 		if sys.version_info.major >= 3:
 			MESSAGE = bytes(MESSAGE, "utf-8")
 		print("Press ctrl+C to quit this application.\n")